+++ renderer/modules/projects/components/accordion/project-creation/project-creation.tsx
@@ -10,7 +10,7 @@ import { StepperContext } from '../../../redux/stepper/stepperContext';
 import { useContext, useEffect } from 'react';
 import { CreateProjectActionData } from '../../../redux/stepper/actions/create-project-action';
 import { CreatorContext } from '../../../redux/creator/creator';
-import { projectCreationProgress } from '../../../../../components/project-execution/projectExecution-ui/ExecutionContants';
+import { projectCreationProgress } from './execution-contants';
 
 export default function ProjectCreation(): JSX.Element {
   const classes = useAccordionStyles();
